splits dlthub cli utils tests
