+++ tests/workspace/test_provider_locations.py
@@ -1,16 +1,9 @@
-import os
-from pathlib import Path
 from typing import Iterator, List
 
 import pytest
 
-from dlt._workspace.cli.dlthub.ai import (
-    ai_secrets_list_command,
-    ai_secrets_view_redacted_command,
-    ai_secrets_update_fragment_command,
-)
 from dlt._workspace.cli.utils import get_provider_locations
-from dlt._workspace.typing import ProviderInfo, ProviderLocationInfo
+from dlt._workspace.typing import ProviderInfo
 
 from tests.workspace.utils import isolated_workspace
 
@@ -137,93 +130,3 @@ def test_get_provider_locations_environ_provider(default_workspace: None) -> Non
     # find providers with no locations (EnvironProvider)
     no_loc_infos = [i for i in infos if len(i.locations) == 0]
     assert len(no_loc_infos) >= 1
-
-
-def test_ai_secrets_list_oss(
-    autouse_test_storage: None,
-    preserve_run_context: None,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """OSS context lists secrets.toml without profiles."""
-    # note: test conftest patches RunContext.initial_providers to use tests/.dlt
-    # so providers point to tests/.dlt, not the workspace's .dlt — that's expected
-    with isolated_workspace("legacy", required="RunContext"):
-        ai_secrets_list_command()
-        output = capsys.readouterr().out
-
-        assert "Secret file locations:" in output
-        # no profile tags in OSS context
-        assert "profile:" not in output
-        # no "not found"
-        assert "not found" not in output.lower()
-
-
-def test_ai_secrets_list_workspace_with_profile(
-    autouse_test_storage: None,
-    preserve_run_context: None,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """Workspace context lists profile-scoped path first, no global."""
-    with isolated_workspace("default", profile="dev"):
-        ai_secrets_list_command()
-        output = capsys.readouterr().out
-
-        assert "Secret file locations:" in output
-        # profile path listed
-        assert "dev.secrets.toml" in output
-        assert "profile: dev" in output
-        # base path listed
-        assert "secrets.toml" in output
-        # no global
-        assert "global" not in output.lower()
-        # no "not found"
-        assert "not found" not in output.lower()
-        # profile-scoped appears before base
-        lines = [line.strip() for line in output.splitlines() if "secrets.toml" in line]
-        assert len(lines) >= 2
-        assert "dev.secrets.toml" in lines[0]
-
-
-def test_ai_secrets_view_redacted_missing_creates_nothing(
-    autouse_test_storage: None,
-    preserve_run_context: None,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """view-redacted on a missing file warns but does not create the file."""
-    with isolated_workspace("empty", profile="dev"):
-        secrets_path = os.path.join(".dlt", "secrets.toml")
-        assert not os.path.exists(secrets_path)
-
-        ai_secrets_view_redacted_command(path=secrets_path)
-        output = capsys.readouterr().out
-
-        assert "not found" in output.lower()
-        # file should NOT be created
-        assert not os.path.exists(secrets_path)
-
-
-def test_ai_secrets_update_fragment_creates_file(
-    autouse_test_storage: None,
-    preserve_run_context: None,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """update-fragment creates secrets file if it doesn't exist."""
-    with isolated_workspace("empty", profile="dev"):
-        secrets_path = os.path.join(".dlt", "secrets.toml")
-        assert not os.path.exists(secrets_path)
-
-        ai_secrets_update_fragment_command(
-            fragment='[sources.my_api]\napi_key = "sk-test-xxx"\n',
-            path=secrets_path,
-        )
-        output = capsys.readouterr().out
-
-        # file was created
-        assert os.path.isfile(secrets_path)
-        # output shows redacted content
-        assert "api_key" in output
-        assert "sk-test-xxx" not in output
-        assert "**" in output
-        # actual file has the real value
-        content = Path(secrets_path).read_text(encoding="utf-8")
-        assert "sk-test-xxx" in content
